requirements.txt
fastmcp>=1.0.0
starlette>=0.27.0
uvicorn>=0.23.0
httpx>=0.24.0 